requirements.txt
langgraph==1.0.5
langchain==1.2.3
langchain-core==1.2.7
langchain-text-splitters==1.1.0
langgraph-checkpoint==3.0.1
langgraph-checkpoint-sqlite>=3.0.0

langchain-google-genai>=2.0.0    # For Google Gemini (default)
langchain-openai>=0.2.0          # For OpenAI GPT-4/3.5 (alternative)

langchain-qdrant==1.1.0
qdrant-client==1.16.2
langchain-huggingface==1.2.0
fastembed==0.7.4
sentence-transformers==5.2.0

PyMuPDF==1.26.6
pymupdf4llm==0.2.9

python-dotenv==1.0.0
pydantic==2.12.5
googlemaps>=4.10.0

gradio==6.3.0

requests==2.32.5
numpy==2.4.1
torch==2.9.1  